land v0.9.1 on HEAD, in preparation for a 1.0.x branch

+++ lustre/tests/multifstat.c
@@ -47,8 +47,9 @@ int main(int argc, char **argv)
         }
 
         if ( st1.st_size != st2.st_size ) {
-                printf("Sizes don't match %ld, %ld\n",
-                       st1.st_size, st2.st_size);
+                printf("Sizes don't match %lu, %lu\n",
+                       (unsigned long)st1.st_size, 
+		       (unsigned long)st2.st_size);
                 return 1;
         }
